#1892692: [Test] Pops up NPE error with invalid values for Quota and Retention Period in Create Web App dialog
