+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/azure/toolkit/intellij/appservice/AppServiceMonitorPanel.form
@@ -195,7 +195,7 @@
             <properties/>
             <border type="none"/>
             <children>
-              <component id="45563" class="com.microsoft.azure.toolkit.intellij.common.IntegerTextField" binding="txtQuota" custom-create="true">
+              <component id="45563" class="com.microsoft.azure.toolkit.intellij.common.AzureIntegerInput" binding="txtQuota" custom-create="true">
                 <constraints>
                   <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
                 </constraints>
@@ -211,7 +211,7 @@
             <properties/>
             <border type="none"/>
             <children>
-              <component id="d3331" class="com.microsoft.azure.toolkit.intellij.common.IntegerTextField" binding="txtRetention" custom-create="true">
+              <component id="d3331" class="com.microsoft.azure.toolkit.intellij.common.AzureIntegerInput" binding="txtRetention" custom-create="true">
                 <constraints>
                   <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
                 </constraints>
