+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/azure/toolkit/intellij/appservice/AppServiceMonitorPanel.java
@@ -9,7 +9,7 @@ import com.intellij.openapi.project.Project;
 import com.intellij.ui.TitledSeparator;
 import com.microsoft.azure.toolkit.intellij.appservice.insights.ApplicationInsightsComboBox;
 import com.microsoft.azure.toolkit.intellij.common.AzureFormPanel;
-import com.microsoft.azure.toolkit.intellij.common.IntegerTextField;
+import com.microsoft.azure.toolkit.intellij.common.AzureIntegerInput;
 import com.microsoft.azure.toolkit.lib.appservice.ApplicationInsightsConfig;
 import com.microsoft.azure.toolkit.lib.appservice.MonitorConfig;
 import com.microsoft.azure.toolkit.lib.appservice.model.DiagnosticConfig;
@@ -43,8 +43,8 @@ public class AppServiceMonitorPanel extends JPanel implements AzureFormPanel<Mon
     private JPanel pnlApplicationLog;
     private TitledSeparator titleApplicationInsights;
     private TitledSeparator titleAppServiceLog;
-    private IntegerTextField txtQuota;
-    private IntegerTextField txtRetention;
+    private AzureIntegerInput txtQuota;
+    private AzureIntegerInput txtRetention;
     private LogLevelComboBox cbLogLevel;
     private JLabel lblApplicationLog;
 
@@ -146,11 +146,11 @@ public class AppServiceMonitorPanel extends JPanel implements AzureFormPanel<Mon
         cbLogLevel = new LogLevelComboBox();
         applicationInsightsComboBox = new ApplicationInsightsComboBox();
 
-        txtQuota = new IntegerTextField();
+        txtQuota = new AzureIntegerInput();
         txtQuota.setMinValue(25);
         txtQuota.setMaxValue(100);
 
-        txtRetention = new IntegerTextField();
+        txtRetention = new AzureIntegerInput();
         txtRetention.setMinValue(0);
         txtRetention.setMaxValue(99999);
         txtRetention.setRequired(false);
