+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/azure/toolkit/intellij/appservice/jfr/DurationPanel.form
@@ -24,7 +24,7 @@
           <toolTipText value=""/>
         </properties>
       </component>
-      <component id="3da98" class="com.microsoft.azure.toolkit.intellij.common.IntegerTextField" binding="textField1" custom-create="true">
+      <component id="3da98" class="com.microsoft.azure.toolkit.intellij.common.AzureIntegerInput" binding="textField1" custom-create="true">
         <constraints>
           <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="2" anchor="8" fill="1" indent="0" use-parent-layout="false">
             <preferred-size width="70" height="-1"/>
